requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
python-multipart
httpx==0.24.1
pydantic-settings
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
greenlet==3.2.3
h11==0.14.0
httpcore==0.17.3
idna==3.10
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic_core==2.33.2
pytesseract==0.3.13
python-dotenv==1.1.0
requests==2.32.4
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
openai
cloud-sql-python-connector[pg8000]
SQLAlchemy>=2.0
google-cloud-storage==2.18.2